add FrontierStiching trigger technique et re-organized examples folder
